+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add helm manifest.
+- Add Upgrade-Schedule-Operator.
 
 
 [Unreleased]: https://github.com/giantswarm/upgrade-schedule-operator/tree/master